resolved PR edits
