+++ client/src/sections/Login/Login.tsx
@@ -1,26 +1,13 @@
 import React, {useState, useCallback} from 'react';
-import {useHistory} from 'react-router-dom';
 import {Button, Card, Heading, ButtonGroup} from '@shopify/polaris';
 import logo from 'icons/logo.png';
 
 import './Login.css';
-import LoginModal from './LoginModal';
+import LoginModal from './LoginModal/LoginModal';
 
 export default function Login() {
-  const history = useHistory();
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [active, setActive] = useState(false);
-
-  const handleChange = useCallback(() => setActive(!active), [active]);
-
-  const handleLogin = () => {
-    history.push('/home');
-  };
-
-  const handleEmailLogin = useCallback((value) => setEmail(value), []);
-
-  const handlePasswordLogin = useCallback((value) => setPassword(value), []);
+  const [showModal, toggleShowModal] = useState(false);
+  const handleChange = useCallback(() => toggleShowModal(!showModal), [showModal]);
 
   return (
     <div className='login'>
@@ -35,8 +22,7 @@ export default function Login() {
           </ButtonGroup>
         </Card>
       </Card>
-      <LoginModal email={email} password={password} active={active} handleChange={handleChange}
-        handleLogin={handleLogin} handleEmailLogin={handleEmailLogin} handlePasswordLogin={handlePasswordLogin}></LoginModal>
+      <LoginModal active={showModal} handleChange={handleChange} />
     </div>
   );
 }
