--- client/src/sections/Login/LoginModal.tsx
@@ -1,40 +1,0 @@
-import {Modal, TextField} from '@shopify/polaris';
-import React from 'react';
-
-export default function LoginModal(props: any) {
-
-  return (
-    <div className='loginModal'>
-      <Modal
-        open={props.active}
-        onClose={props.handleChange}
-        title="Login"
-        primaryAction={{
-          content: 'Login',
-          onAction: props.handleLogin,
-        }}
-        secondaryActions={[
-          {
-            content: 'Cancel',
-            onAction: props.handleChange,
-          },
-        ]}
-      >
-        <Modal.Section>
-          <TextField
-            value={props.email}
-            onChange={props.handleEmailLogin}
-            label="Email"
-            type="email"
-          />
-          <TextField
-            value={props.password}
-            onChange={props.handlePasswordLogin}
-            label="Password"
-            type="password"
-          />
-        </Modal.Section>
-      </Modal>
-    </div>
-  );
-}
