+++ client/src/sections/Login/components.ts
@@ -1,0 +1,1 @@
+export {default as LoginModal} from './LoginModal'
